remove debug log statement in EXECUTOR shutdown callback
Using the signals module the library tries to gracefully
shutdown the ThreadPoolExecutor with a callback. For us
this causes some issues with the logger module warning
that you're writing to write to a log file after
the file is closed. With these signal callbacks
you can't control the shutdown order.